# [CPA] Additional optimizations to graph constructions