## What does this PR do?
Improvements
1. Instead of using df.apply() we use zip to speed up object construction, see [this post](https://stackoverflow.com/questions/76502222/what-is-the-fastest-way-to-iterate-through-a-pandas-dataframe-to-create-custom-o?fbclid=IwZXh0bgNhZW0CMTEAAR0DwfvGhHi9mTIFXxJG4_DkNmFhBVOQziOcsCJbZjcpRVl7rqVGaZ4hB7A_aem_hd1QvPwJ3AU5R2Hg5ojt9g)
1. Use itertuples instead of iterrows while processing gpu kernels.
1. Cache the is blocking call check to avoid dataframe loc lookup

## Testing
  pytest tests/test_critical_path_analysis.py

## Timing improvements
Before: CPGraph construction took 16.80 seconds
   After: CPGraph construction took 6.35 seconds

Function level breakdown
Before
```
Function _create_event_nodes Took 5.3051 seconds
Function _construct_graph_from_call_stacks Took 6.3984 seconds
Function _get_cuda_event_record_df Took 0.4013 seconds
Function _get_cuda_stream_wait_event_df Took 0.5246 seconds
Function _construct_graph_from_kernels Took 5.0919 seconds
```
After
```
Function _create_event_nodes Took 0.4935 seconds
Function _construct_graph_from_call_stacks Took 3.6381 seconds
Function _get_cuda_event_record_df Took 0.4002 seconds
Function _get_cuda_stream_wait_event_df Took 0.5362 seconds
Function _construct_graph_from_kernels Took 2.1994 seconds
```


## Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [x] N/A
- [x] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [x] N/A